+++ lib/rubocop/cop/sidekiq/no_nil_return.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+module RuboCop
+  module Cop
+    module Sidekiq
+      # Ensure workers avoid returning early
+      #
+      # # bad
+      # def perform
+      #   return if condition
+      #   ...
+      # end
+      #
+      # # good
+      # def perform
+      #   # TDB, idea would be `raise SilentError if condition`
+      # end
+
+      class NoNilReturn < Base
+        MSG = 'Avoid using early nil return in workers.'
+
+        def on_return(node)
+          add_offense(node) if node.arguments.first&.nil_type? || node.arguments.empty?
+        end
+      end
+    end
+  end
+end